（大塚）：Android Ver1.1.0の内容

+++ Assets/DataSaver.cs
@@ -47,9 +47,9 @@ public class DataSaver : MonoBehaviour {
         switch (pedocountsetmode)
         {
             case PEDOCOUNTSETMODE.ADDCITIVE:
-                int t_pedocount = 0;
-                int.TryParse(newPedocount, out t_pedocount);
-                pedocount = initialPedoCount + t_pedocount;
+                float t_pedocount = 0;
+                float.TryParse(newPedocount, out t_pedocount);
+                pedocount = initialPedoCount + (int)t_pedocount;
                 break;
             case PEDOCOUNTSETMODE.OVERWRITE:
                 int.TryParse(newPedocount, out pedocount);
